adds solution for binary search exercise
