+++ exercises/algorithms/src/main/java/com/cbfacademy/App.java
@@ -1,13 +1,13 @@
 package com.cbfacademy;
 
-import com.cbfacademy.search.SearchAlgoExercise;
+import com.cbfacademy.search.BinarySearch;
 
 public class App {
     public static void main(String[] args) {
-        SearchAlgoExercise searchAlgoExercise = new SearchAlgoExercise();
+        BinarySearch binarySearch = new BinarySearch();
         int[] sortedArray = { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 };
         int target = 12;
 
-        System.out.println(searchAlgoExercise.binarySearch(sortedArray, target));
+        System.out.println(binarySearch.search(sortedArray, target));
     }
 }
